feat: extract text renderer to a file
