+++ packages/nextjs-contentful-renderer/src/MdxRenderer/renderers/index.ts
@@ -1,0 +1,1 @@
+export * from './text';
